Move DateProvider and Scheduler to api/common
We need them on the worker side too e.g. for sleep reconnect in
EventBusClient.
