+++ src/api/common/DateProvider.ts
@@ -1,0 +1,5 @@
+export interface DateProvider {
+    now(): number
+
+    timeZone(): string
+}
