+++ src/api/main/MainLocator.ts
@@ -16,7 +16,7 @@ import type {DeferredObject} from "@tutao/tutanota-utils"
 import {defer, downcast} from "@tutao/tutanota-utils"
 import {ProgressTracker} from "./ProgressTracker"
 import {MinimizedMailEditorViewModel} from "../../mail/model/MinimizedMailEditorViewModel"
-import {SchedulerImpl} from "../../misc/Scheduler"
+import {SchedulerImpl} from "../common/utils/Scheduler"
 import type {ICredentialsProvider} from "../../misc/credentials/CredentialsProvider"
 import {createCredentialsProvider} from "../../misc/credentials/CredentialsProviderFactory"
 import type {LoginFacade} from "../worker/facades/LoginFacade"
