+++ test/api/worker/SuspensionHandlerTest.ts
@@ -2,7 +2,7 @@ import o from "ospec"
 import {SuspensionHandler} from "../../../src/api/worker/SuspensionHandler"
 import {deferWithHandler, downcast} from "@tutao/tutanota-utils"
 import type {WorkerImpl} from "../../../src/api/worker/WorkerImpl"
-import type {SystemTimeout} from "../../../src/misc/Scheduler"
+import type {SystemTimeout} from "../../../src/api/common/utils/Scheduler"
 o.spec("SuspensionHandler test", () => {
     let suspensionHandler
     let workerMock: WorkerImpl
